# docs/openapi.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/VivekVinushanth/DataIngestion/1.0.0
info:
  description: API for ingesting profile events and personality data
  version: "1.0.0"
  title: Data Ingestion API
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
paths:
  /{perma_id}/{app_id}/event:
    post:
      summary: Push profile events
      operationId: addEvent
      description: Adds events to the profile
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the profile
        - in: path
          name: app_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: Event added successfully
        '400':
          description: Invalid event object
  /{perma_id}/{app_id}/events:
    post:
      summary: Push user events
      operationId: addEvents
      description: Adds events in bulk to the user profile
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the user
        - in: path
          name: app_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: Event added successfully
        '400':
          description: Invalid event object

    get:
      summary: Fetch events emitted by a user
      operationId: getUserEvents
      description: Retrieve events emitted by a specific user
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the profile
        - in: path
          name: app_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the application
        - in: query
          name: searchString
          schema:
            type: string
          description: Optional search filter
        - in: query
          name: offset
          schema:
            type: integer
            format: int
            minimum: 0
          description: Number of records to skip for pagination
        - in: query
          name: limit
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 50
          description: Maximum number of records to return
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Bad input parameter
  /{perma_id}/{app_id}/events/{event_id}:
    get:
      summary: Fetch specific event emitted by a user
      operationId: getUserEvent
      description: Retrieve a specific event emitted by a specific profile
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the profile
        - in: path
          name: app_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the application
        - in: path
          name: event_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the event
        - in: query
          name: searchString
          schema:
            type: string
          description: Optional search filter
        - in: query
          name: offset
          schema:
            type: integer
            format: int
            minimum: 0
          description: Number of records to skip for pagination
        - in: query
          name: limit
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 50
          description: Maximum number of records to return
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Bad input parameter
...
  /{perma_id}/consent:
    get:
      summary: Fetch profileconsent
      operationId: getConsentedApps
      description: Retrieve consented applications to collect data
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the user
      responses:
        '200':
          description: Fetch list of consented apps to collect data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consent'

        '400':
          description: Bad input parameter
  /{perma_id}/consent/collect:
    post:
      summary: provide consent to collect data
      operationId: giveConsentToCollect
      description: Capture profile consent to collect data
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uuid
      responses:
        '201':
          description: Alias successfully updated
        '400':
          description: Invalid alias request

    get:
      summary: Fetch applications users has given consent to collect
      operationId: getConsentedAppsToCollect
      description: Retrieve consented applications to collect data
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the user
      responses:
        '200':
          description: Fetch list of consented apps to collect data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid

        '400':
          description: Bad input parameter
  /{perma_id}/consent/share:
    post:
      summary: provide consent to share profile
      operationId: giveConsentToShare
      description: Capture profile consent to share data
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uuid
      responses:
        '201':
          description: Alias successfully updated
        '400':
          description: Invalid alias request

    get:
      summary: Fetch applications users has given consent to share
      operationId: getConsentedAppsToShare
      description: Retrieve consented applications to share profile data
      parameters:
        - in: path
          name: perma_id
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the user
      responses:
        '200':
          description: Fetch alias of a profile
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid

        '400':
          description: Bad input parameter
components:
  schemas:
    Event:
      type: object
      required:
        - perma_id
        - app_id
        - event_type
        - event_name
        - event_id
        - event_timestamp
        - context
      properties:
        event_type:
          type: string
          enum: [Identify, Track, Page]
          description: Type of the event
        event_name:
          type: string
          example: "product_viewed"
          description: Name of the event
        event_id:
          type: string
          format: uuid
          description: Unique ID for the event
        app_id:
          type: string
          format: uuid
          description: Unique ID for the application
        event_timestamp:
          type: string
          format: date-time
          description: Time at which the event occurred
        context:
          type: object
          description: Device and session information
        locale:
          type: string
          example: "en-US"
          description: User's language and regional settings
        properties:
          oneOf:
            - $ref: '#/components/schemas/PageEvent'
            - $ref: '#/components/schemas/TrackEvent'
            - $ref: '#/components/schemas/IdentifyEvent'
          discriminator:
            propertyName: event_type
            mapping:
              Page: "#/components/schemas/PageEvent"
              Track: "#/components/schemas/TrackEvent"
              Identify: "#/components/schemas/IdentifyEvent"
    PageEvent:
      type: object
      description: Event properties for page interactions
      properties:
        page:
          type: object
          properties:
            url:
              type: string
              description: Full URL of the page
            path:
              type: string
              description: Page path without domain
            referrer:
              type: string
              description: URL of the previous page
            title:
              type: string
              description: Title of the page
            search:
              type: string
              description: Query parameters from the URL
            page_category:
              type: string
              description: Logical category of the page
            page_type:
              type: string
              description: Type of page (e.g., landing_page, blog)
            content_type:
              type: string
              description: Type of content (e.g., article, video)
            scroll_depth:
              type: string
              description: Percentage of page scrolled
            time_on_page:
              type: integer
              description: Time spent on the page in seconds
            previous_page:
              type: string
              description: Identifier for the previous page
        utm:
          type: object
          properties:
            source:
              type: string
              description: Traffic source (e.g., google, facebook)
            medium:
              type: string
              description: Marketing medium (e.g., email, social)
            campaign:
              type: string
              description: Campaign name
        engagement:
          type: object
          properties:
            engagement_score:
              type: number
              description: Custom engagement score
            interactive_elements:
              type: array
              items:
                type: string
              description: List of interactive elements clicked
    TrackEvent:
      type: object
      description: Event properties for tracking user interactions
      properties:
        action:
          type: string
          description: The specific action performed (click, scroll)
        object_type:
          type: string
          description: The type of object interacted with (button, product)
        object_id:
          type: string
          description: Unique identifier for the interacted object
        object_name:
          type: string
          description: Human-readable name of the object
        value:
          type: number
          description: A numeric value associated with the event
        label:
          type: string
          description: Additional label for categorization
        source:
          type: string
          description: Source of the interaction (website, mobile app)
        url:
          type: string
          description: URL where the event occurred
        referrer:
          type: string
          description: URL of the referring page
    IdentifyEvent:
      type: object
      description: Event properties for user identity tracking
      properties:
        user_id:
          type: string
          format: uuid
          description: Unique identifier for the user
        traits:
          type: object
          description: Custom user attributes
          additionalProperties: true
...
    Consent:
      type: object
      required:
        - collect
        - share
      properties:
        collect:
          type: array
          items:
            type: string
            format: uuid
        share:
          type: array
          items:
            type: string
            format: uuid